Stream based command used for location
RxDart added, Stream based Command used to fetch location.  UI was reworked for more symmetry and a better look.  lastKnownLocation method is not used to get location anymore because it is inconsistent.  CurrentLocation method is now used in its place.  

New Object made to combine the two boolean Obervables so that they can disable both the location and weather commands based on more complex logic.